requirements.txt
fastapi
uvicorn
bcrypt
requests
numpy
faiss-cpu
sentence-transformers
llama-cpp-python
gpt4all
llama-index-core
llama-index-vector-stores-faiss
llama-index-embeddings-openai
openai
python-dotenv
google-generativeai
psycopg2-binary
llama-index-vector-stores-postgres
python-jose